exp 7

+++ scripts/clear_results.sh
@@ -3,6 +3,12 @@ rm datasets/icub_alone/dataset_optical_flow.npy
 rm datasets/only_ball/dataset_optical_flow.npy
 rm datasets/icub_and_ball/dataset_optical_flow.npy
 rm datasets/icub_and_many_balls/dataset_optical_flow.npy
-echo "Deleting old results and models"
-rm -r experiments/
-echo "done!"
+
+read -p "Do you want to remove the experiments folder? " -n 1 -r
+echo    # (optional) move to a new line
+if [[ $REPLY =~ ^[Yy]$ ]]
+then
+    echo "Deleting old results and models"
+    rm -r experiments/
+    echo "removed!"
+fi
